requirements.txt
alembic==1.10.2
attrs==22.2.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
flake8==6.0.0
flake8-commas==2.1.0
greenlet==2.0.2
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pytest==7.2.2
pytest-env==0.8.1
pytest-lazy-fixture==0.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
rich==12.6.0
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==2.0.6
terminaltables==3.1.10
typer==0.7.0
typing_extensions==4.5.0
